Move the mixins into Concerns
This makes the app a bit more standard and should fix the require issues
